Bug 1134268 - Part 2 - Fixup TelemetryEnvironment shutdown if the module wasnt initialized. r=vladan

+++ toolkit/components/telemetry/TelemetryEnvironment.jsm
@@ -261,8 +261,12 @@ this.TelemetryEnvironment = {
    */
   shutdown: Task.async(function* () {
     if (this._shutdown) {
-      this._log.error("shutdown - Already shut down");
-      throw new Error("Already shut down");
+      if (this._log) {
+        this._log.error("shutdown - Already shut down");
+      } else {
+        Cu.reportError("TelemetryEnvironment.shutdown - Already shut down");
+      }
+      return;
     }
 
     this._log.trace("shutdown");
